Adicionando metodos para visualizar um usuario específico por ID e editar um usuario específico por ID

+++ src/main/java/murilo/api_autenticacao/controllers/AdminController.java
@@ -22,13 +22,44 @@ public class AdminController {
     @GetMapping("/users")
     public ResponseEntity<List<UserResponseDTO>> getAllUsers() {
         List<User> users = userRepository.findAll();
-        List<UserResponseDTO> userDtos = users.stream().map(this::convertToDto).collect(Collectors.toList());
+        List<UserResponseDTO> userDtos = users.stream()
+                .map(this::convertToDto)
+                .collect(Collectors.toList());
         return ResponseEntity.ok(userDtos);
+    }
+
+    @GetMapping("/users/{id}")
+    public ResponseEntity<UserResponseDTO> getUserById(@PathVariable Long id) {
+        return userRepository.findById(id)
+                .map(this::convertToDto)
+                .map(ResponseEntity::ok)
+                .orElse(ResponseEntity.notFound().build());
+    }
+
+    @PutMapping("/users/{id}")
+    public ResponseEntity<UserResponseDTO> updateUser(@PathVariable Long id, @RequestBody UserResponseDTO userData) {
+        return userRepository.findById(id)
+                .map(userToUpdate -> {
+
+                    userToUpdate.setName(userData.getName());
+                    userToUpdate.setEmail(userData.getEmail());
+                    userToUpdate.setRole(userData.getRole());
+
+                    User updatedUser = userRepository.save(userToUpdate);
+
+                    return ResponseEntity.ok(this.convertToDto(updatedUser));
+                })
+                .orElse(ResponseEntity.notFound().build());
     }
 
     @DeleteMapping("/users/{id}")
     public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
+        if (!userRepository.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
+
         userRepository.deleteById(id);
+
         return ResponseEntity.noContent().build();
     }
